Cover letters complete
Complete cover letter import with example. I'm a little embarrassed about the content but seeing it has spurred me to change my template to have a better intro and career alignment sections. I kind of just explained who I was, not emphasized how I match. The list would be tailored to the company along with the closing paragraph but this could be better. If I were to interview at Spreedly again, I'd choose a different cover letter.