Elevator pitch complete
The elevator pitch directory is kind of the one-stop interview emporium even though it's meant to just be a 90 second pitch to who I am. I feel like all of that is "sticky" though and relevant in the same situations. My resume summary and examples from LinkedIn or personal websites is applicable to a degree as well. I don't have the best categorization skills but I also don't want a billion directories.