Job Opportunities | Correspondence
Basic correspondence structure. This may not be exhaustive or follow anyone else's workflows but it's meant to be a starting point.
I firmly believe correspondence should be very personal but these templates could easily be a guide, either a checklist or a very specific set of instructions to follow for your end result.