Housekeeping | Action Plan Opportunities
Markdown linting cleanup for 2021 files
Decided to split by seasons to dupe files instead of have things like elixir-chat--2022 and 2021 in the same directory. Ideally these would be versioned instead of yearly which gets messy as in my example, I did a late 2021-early 2022 stint.
Add late 2022 stubs.
Action plans separate from the opportunities feels a little messy as I may work better for things that are more self contained. It is an action plan but that may not apply to all relevant files.